Move the release creation to a separate process
It should avoid any stalled jobs.
