+++ api/lib/infrastructure/scheduled-jobs/check-urls-job.js
@@ -2,6 +2,7 @@ const createQueue = require('./create-queue');
 const config = require('../../config');
 
 const queue = createQueue('check-urls-queue');
+queue.process(__dirname + '/check-urls-job-processor.js');
 
 const checkUrlsJobOptions = {
   attempts: config.scheduledJobs.attempts,
@@ -9,8 +10,6 @@ const checkUrlsJobOptions = {
   removeOnComplete: true,
   removeOnFail: 1,
 };
-
-queue.process(__dirname + '/check-urls-job-processor.js');
 
 function start() {
   queue.add({}, checkUrlsJobOptions);
